pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "tuesday"
version = "0.1.0"
description = ""
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
Pillow = "^11.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
